Move methods to agenda
