+++ js/ui/scheduler/workspaces/ui.scheduler.agenda.js
@@ -15,7 +15,7 @@ import {
     GROUP_ROW_CLASS,
     GROUP_HEADER_CONTENT_CLASS,
 } from '../classes';
-import { getPathToLeaf } from '../resources/utils';
+import { createReducedResourcesTree, getPathToLeaf } from '../resources/utils';
 import { calculateStartViewDate } from '../../../renovation/ui/scheduler/view_model/to_test/views/utils/agenda';
 import { formatWeekday, getVerticalGroupCountClass } from '../../../renovation/ui/scheduler/view_model/to_test/views/utils/base';
 import { VIEWS } from '../constants';
@@ -238,8 +238,15 @@ class SchedulerAgenda extends WorkSpace {
 
     _makeGroupRows() {
         const { filteredItems } = this.invoke('getAppointmentDataProvider'); // TODO refactoring
+
         const resourceManager = this.option('resourceManager');
-        const tree = resourceManager.createReducedResourcesTree(filteredItems); // TODO refactoring
+
+        const tree = createReducedResourcesTree(
+            resourceManager.loadedResources,
+            (field, action) => resourceManager.getDataAccessors(field, action),
+            filteredItems
+        ); // TODO refactoring
+
         const cellTemplate = this.option('resourceCellTemplate');
         const getGroupHeaderContentClass = GROUP_HEADER_CONTENT_CLASS;
         const cellTemplates = [];
